Make deck cards clickable
