+++ lib/app/app.dart
@@ -1,5 +1,4 @@
 import 'package:flash/config/theme.dart';
-import 'package:flash/flashcards/flashcards_view.dart';
 import 'package:flash/navigation/navigation_view.dart';
 import 'package:flutter/material.dart';
 
@@ -12,8 +11,7 @@ class App extends StatelessWidget {
       title: 'Flash',
       debugShowCheckedModeBanner: false,
       theme: kLightTheme,
-      // home: const NavigationView(),
-      home: const FlashcardsView(),
+      home: const NavigationView(),
     );
   }
 }
